Moved common methods in GameIO child classes into GameIO proper.
Changed GUI master class to implement GameIO now that it is non-empty.

Cleaned up interfaces to match new changes.
