+++ src/main/java/usecases/IOInterfaces/GoFishIO.java
@@ -16,13 +16,6 @@ public interface GoFishIO extends GameIO {
     void changePlayer(String username);
 
     /**
-     * This method should send a popup to the user containing a <code>message</code>.
-     *
-     * @param message a string that is to be sent to the user
-     */
-    void sendPopup(String message);
-
-    /**
      * This method should display the user's hand.
      *
      * @param hand a string representation of a hand of cards.
@@ -46,16 +39,4 @@ public interface GoFishIO extends GameIO {
      * @return a string representation of the username selected by the user
      */
     String getPlayerUsername(String currPlayer, List<String> usernames);
-
-    /**
-     * This method should display a <code>message</code> to the user and then close the GUI.
-     *
-     * @param message a string that is to be sent to the user.
-     */
-    void closeMessage(String message);
-
-    /**
-     * This method should close the GUI when called.
-     */
-    void close();
 }
