+++ src/main/java/usecases/IOInterfaces/WarIO.java
@@ -29,22 +29,4 @@ public interface WarIO extends GameIO {
      */
     void stall();
 
-    /**
-     * This method should send a popup to the user containing a <code>message</code>.
-     *
-     * @param message a string that is to be sent to the user
-     */
-    void sendPopup(String message);
-
-    /**
-     * This method should close the GUI when called.
-     */
-    void close();
-
-    /**
-     * This method should display a <code>message</code> to the user and then close the GUI.
-     *
-     * @param message a string that is to be sent to the user.
-     */
-    void closeMessage(String message);
 }
