+++ src/test/java/usecases/GameTemplateTest.java
@@ -37,6 +37,32 @@ class GameTemplateTest {
 
     static class FakeGameIO implements GameIO {
 
+        /**
+         * This method should send a popup to the user containing a <code>message</code>.
+         *
+         * @param message a string that is to be sent to the user
+         */
+        @Override
+        public void sendPopup(String message) {
+        }
+
+        /**
+         * This method should display a <code>message</code> to the user and then close the GUI.
+         *
+         * @param message a string that is to be sent to the user.
+         */
+        @Override
+        public void closeMessage(String message) {
+
+        }
+
+        /**
+         * This method should close the GUI when called.
+         */
+        @Override
+        public void close() {
+
+        }
     }
 
     static class FakeGame extends GameTemplate {
